eatout_app/requirements/dev.txt
-r common.txt
django-jet==1.0.6
Pillow==4.3.0